requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-admin-bootstrapped==2.5.7
django-appconf==1.0.1
django-imagekit==3.3
django-nested-admin==3.0.8
django-toolbelt==0.0.1
djangorestframework==3.3.3
gunicorn==19.4.5
pilkit==1.1.13
Pillow==3.3.1
psycopg2==2.6.1
python-monkey-business==1.0.0
requests==2.7.0
requests-futures==0.9.5
six==1.10.0
South==1.0.2
static3==0.6.1
whitenoise==3.0
